Fix line splitting when parsing status of Draytek devices
The status output contains lines with two items, which are detected by
having two colons in them. Both are separated by splitting where the
longest consecutive whitespace is (after some pre-processing to remove
some whitespace). However, the splitting could be done before the first
colon, which lead to an infinite loop, as the two substrings are again
processed by the same function.

To fix this, splitting at the longest whitespace is now only performed
in the substring within the two colons.

This issue occured when the first of the items had an empty value, for
example when "Running Mode" is empty (which is true before sync).