Support entire valid value range for Hlog graphs
This extends the Hlog graph to support values from 0 and 6 dB, so that
the entire possible value range is shown.

The graph had to be slightly extended to support values above 5, which
also happens to make the legend items more closely align the pixel grid
at the default graph size.