Fix recently introduced error in Hlog path generation
Each line segment of the path segment is ended with a point that is
added on transition from a valid to an invalid data point. Due to an
error introduced with commit af20b29d7c4b7c4a6b88c0327b1c64cc1b21b704,
this point was added for every pair of valid data items.