Change bins data structure to more closely reflect actual data
Except for the bitloading data, the bins data is often not available for
every bin, but only for groups of bins. The group size may differ for
downstream and upstream data of the same kind (e.g. because one of them
is limited to a smaller number of bins as with VDSL2 profile 35b, or
because locally measured data is available in higher resolution).

Thus, all the bins data is now stored separated into downstream and
upstream values, and the grouping is preserved as in the data parsed
from the modem (or reconstructed in case of Broadcom). Band plan data
is now stored explicitly, if available.

This change allows clients to better interpret the data, for example
graphs can be drawn more accurately. The split into downstream and
upstream will also be useful if G.fast support is added in the future.