Add unified client constructor
This adds a unified constructor to the main package, which takes a
client type and configuration options, and can create a client for any
supported device type.

In order to do this, there is a client type registry. Clients register
themselves in their init function. Thus, the consumer needs to import
the relevant client package to use it.