Add support for device-specific options
These options are specified in the unified configuration using a map,
so there does not need to be a field for configuration options that are
only used for one or few device types.

ClientDesc contains descriptions of supported options, so that
applications can validate options and provide help to the user.