Handle Telnet server that re-writes prompt after receiving command
This detects a carriage return at the start of the read data (which
means that the prompt was overwritten), and continues reading data until
the next prompt in that case.

The particular server that requires this sometimes sends a carriage
return, the prompt, and the ANSI control sequence "Erase in Display"
before echoing the command.

A better solution would be to first find the echoed command and use it
as start delimiter for the command output. However, we don't know if
server side echo is actually in use, so this could for break some
servers. This information could be read from the server, but the Telnet
client currently does not expose the data.