Read minimum error-free throughput
This value is a good indicator for transmission errors on the line, as
it shows the impact on bandwidth due to retransmissions.

This value is available for Broadcom and Lantiq devices. While it should
be useful in theory, it is not entirely clear if that applies to the
actual reported values. At least for Broadcom it seems that EFTR_min is
reported instead of MINEFTR, which only provides an instantaneous value.