Read vectoring state
This is supported at least partially for all device types. For Lantiq,
there is full support. For Draytek there is no known way to detect
vectoring friendly state. For Broadcom, there is only a single value
that does not tell anything about friendly mode or directions, so only
the downstream value is set if vectoring is in use.