Add FRITZ!Box client
This adds a client for FRITZ!Box devices. It has been tested with
firmware versions 6.86 and 7.28 on Lantiq-based devices.

Several data sources are used. Most of the information is available from
the web interface. Some error counters and the transmit power are loaded
from TR-064, which needs to be enabled for this client. Optionally, data
for the QLN/Hlog graphs, band information and retransmission counters
may be read from support data.