Add initial support for MediaTek modems
For now, only the subset of information which is easily available from
procfs is supported.

There are other ways to get more detailed information, such as the
"adslphxcmd" and "wan" commands. However, the former doesn't work on the
tested device (the kernel log shows "(get_adsl_dev_ops) p is NULL").
And the output of the "wan" ("tcci") command is printed to the kernel
log instead of stdout, making it difficult to parse (especially for long
output, such as bitloading/SNR/QLN/Hlog).