Read more detailed information for MediaTek devices
This adds support for various additional data:
- Error seconds and retransmission counters
- Fixed INP values and interleaving delay
- Pilot tones and band borders
- QLN and Hlog graphs

The additional data is taken from the "wan" command. As its output is
printed to the kernel log instead of stdout, this uses a hack using
dmesg and sed to read the result.

As it seems that part of the procfs output is specific to Asus devices,
some already supported values are now additionally read from the "wan"
command as well.