Add support for Sagemcom devices
This adds support for Sagemcom devices which expose data in a format
that is similar to the TR-181 device data model.

For now, this has been tested to work with a Speedport Pro. Additional
preparations to also support Broadcom modems were made based on a data
dump from a Sagemcom F@st 5350GV.

There are some quality issues with the data that cannot be fixed on the
client side, such as the "Actual Net Data Rate" not being reported, or
the VDSL2 profile being incorrectly reported.