Fix handling of bins data quirks for Sagemcom devices
Some of the reported values contain excess data at the end. The existing
code for handling this assumed that the excess data was either the fixed
value "139" or a duplicate of the actual value itself.

However, the excess data actually consists of other values from the
TestParams object. This replaces the implementation by a more robust one
that can handle arbitrary excess data, as long as the value matches the
expected format.

This also extends a comment to better explain the check for duplicate
downstream data in the upstream values.