Swap downstream/upstream rtx-tx values for Lantiq devices
Looking at the reported numbers, it seems that RTX-TX the direction is
currently swapped for the rtx-tx values (the increase of rtx-tx must be
lower than the sum of the corresponding rtx-c and rtx-uc values).

This change also matches the names of the reported values (i.e. "Tx" vs
"Rx" for the other retransmission counters).