Add custom error types for authentication and connection failure
This allows to distinguish a few special error cases:

AuthenticationError is returned when authentication fails, and provides
a non-zero WaitTime field when further login attempts are going to be
blocked for some time.

ConnectionError should be returned whenever there is an error that the
client will not recover from (for example when a SSH/Telnet connection
is interrupted, or when a session is no longer valid).