Make client types unexported and return dsl.Client in constructors
There is no need to actually export the individual client types.

As a result of this change, dsl.NewClient now returns untyped nil in
case of errors, instead of the typed nil values it returned previously.

Typed nil values are an issue as they are not obvious and there is no
easy way to detect them. Because of this, they could easily cause users
of the library to unintentionally do method calls on nil objects.