Clean up mode and state parsing
Over time, support for more and more devices was added to the state
parser, and at this point it has become unmaintainable. It is not clear
which parts are required for which devices, and any changes are prone to
cause regressions for existing devices. This moves most of it to the
individual device clients, and only keeps a TR-06X state parser in the
internal helpers package.

The mode parser is kept as is, as it is not really device-specific.
However, it is also moved to an internal package, as there is no need to
expose it in the library API.