Use smaller brotli window size in web interface
This reduces the brotli window size to 256 KiB to reduce per-client
memory usage. This doesn't affect compression performance in a
meaningful way, even for devices that provide very detailed bins data
(e.g. for Broadcom with VDSL2 profile 35b, the typical data size is
about 1/6 of the window size now).