Add basic graphical user interface
For now, this is just a small wrapper around the web interface, and
should be considered an experimental feature. As cgo is required for
building, it has to be enabled with the build tag "gui".

Configuration still needs to be done via command-line arguments. The
save button from the web interface is hidden, as there is no cross-
platform support for downloads in the webview package.