Add types for device-specific options
Options can now be of string or bool type. This allows to use a checkbox
instead of a text input field in the GUI for boolean options. Additional
types can be added in the future as needed.

This is a breaking change, as OptionDescriptions field in ClientDesc is
replaced by a new map called Options which maps option names to a struct
instead of just a string description.