Add errors history
This allows to log the increase in error counter values. By default,
data for up to 24 hours is kept, grouped in bins of 5 minutes.

In some cases, the original counter values are likely to be incorrect.
The code attempts to detect these situations, such as when the line is
(or recently was) not in showtime state, or a long time has elapsed
since the last measurement.