Try to determine the current period for errors history more accurately
Since updates are not synchronized with history periods, it is possible
for an update to arrive shortly after a new period has started. In this
case, it makes more sense to update the previous period before creating
a new one. This change tries to accomplish that by picking the period
with the greatest overlap to the actual interval between updates.