Build Linux binaries for multiple architectures in GitHub workflow
This switches to running in a docker container, as the Ubuntu runner
from GitHub has broken multi-arch support due to use of packages from
unofficial package sources. A side effect of this is that it allows to
switch to a supported runner while still building on the older Ubuntu
release for more compatible binaries.

Some dependencies are installed in each build step instead of up front,
because the libgtk-3-dev package can't be installed for multiple
architectures at the same time.