Simplify bins parsing for Fritzbox support data and fix group size
For "HLOG Array" and "HLOG Array", assume that only downstream data is
provided. This allows to remove a large chunk of parsing code.

Also fix group size calculation to take into account the bands data.
This is not entirely correct, but tries to do the best with the
available data (per the specification, group size is derived from the
SUPPORTEDCARRIERS set, but the bands data is based on the MEDLEY set
with coarse granularity).