Fix swapped error counters for Lantiq devices in CO mode
The error counters are reported for near and far end, so the mapping to
downstream and upstream depends on whether the device runs in CPE or CO
mode.

Detect CO mode based on version numbers to handle this. As there isn't
any detailed documentation on version numbers this may not always be
correct, but it is better than nothing.