Support different subtrees in Lancom client
On devices with multiple modems or an ADSL modem, the data may only be
accessible from other subtrees than the one currently used.

The new "Subtree" option allows to specify the LCOS menu subtree to use
(which is then internally converted to the matching SNMP OID). If the
option is not specified, the subtrees for VDSL and ADSL are tried.