Provide more detailed data for OLR counters (bitswap/SRA)
Replace BoolValue by a new OLRValue type, which provides the number of
executed changes in addition to the enabled state. For now, just the
part which is already supported by the existing clients is populated.

This changes the type of some fields in the models.Status struct, and
also renames the bitswap fields, so this is an incompatible change for
users of the library.