Always apply zero detection when parsing bins in Lantiq client
The issue of all zeros being reported is not actually specific to Vinax.
It also happens at least with ADSL2+ on VR9, and the issue also exists
for SNR.

Remove the special casing of Vinax devices, and add a more generic
solution that detects when all values are zero. Since this no longer
applies to individual values, this should be relatively safe to do in
all cases.