WIP: try to fix empty output in Lantiq client for Fritzbox devices
The dsl_pipe command on Fritzbox devices doesn't output anything by
default, unless a debug level option is changed. Try to do that when
no output is detected as result of running the "vig" command.

TODO: Ths index of the debug level option may not be the same on all
devices / firmware versions. Only tested on 7412 so far.