Report more accurate connection state
The state is reported at different detail level depending on the device,
and the current state values are not a good match for this. For example,
most devices to not distinguish between the idle and silent states, and
not all devices distinguish different phases of the initialization
procedure. This means that some clients are forced to report a more
detailed state than the actual data allows for.

Define a few general states (down, init, showtime, error), and treat any
additional states as more detailed variants of those. This should allow
to always represent the state and still provide detailed information
where available.