Fix handling of invalid SNR values in bins history
Currently, invalid values and zero values are not handled correctly.
This can result in spikes when a bin that was initially valid becomes
invalid, and in missing data when a bin that was initially invalid
becomes valid. The actual behaviour might differ depending on how the
device client handles invalid values.

Fix this by initially setting the min/max data to values outside the
valid range, and checking if values are valid during update.