Fix handling of SNR group size and invalid values in DrayTek client
The "adsl showbins" command returns zero for invalid values, so replace
any zero values by the invalid value -32.5.

However, there is also another command, "adsl status snr", which allows
to distinguish invalid values. It also reports the group size of SNR
data (so it is not necessary to guess it as with the showbins command).
Add support for that command, but keep support for the currently used
"adsl showbins" command (which is ready anyway for bit allocation).