Improve handling of invalid SNR values in MediaTek client
Invalid values are reported as zero, so initialize the data with the
invalid value -32.5, and only treat zero values as invalid while reading
the actual values.

The command "wan vdsl2 show pmdtestparam snr" would also provide raw SNR
values (with distinguishable invalid values). However, as its output is
only accessible via dmesg and thus potentially unrealiable, it doesn't
seem to be worth it to use it.