Add support for optional legend in graphs
If enabled, a description of the graph content with color legend will be
shown below the graph. For now, any graph size below the default is not
supported, and the labels may be cut off otherwise.

For users of the library needing more control, the graphs package also
provides functions to access the raw legend data (title and color map).