Add more space for graph y-axis labels and increase default graph size
The default graph size is increased by 6 pixels, all of which is used
for y-axis labels. This allows to get rid of the transform that was
previously used to squeeze long labels into the available space.

The space reserved for the labels is calculated based on the largest
expected number of digits, which can be adjusted for individual graphs
in the future. For now, all graphs use the same value to keep visual
consistency.