Dynamically calculate step size for graph axis labels
Instead of a single step size, the graph spec now contains a list of
allowed values. The actual value will then be chosen automatically
based on the graph size. The label start position is also adjusted to
match a multiple of the step size.

For small graphs, the step size is determined based on the available
space directly. With increasing graph size, additional spacing is added,
to make the graphs appear less crowded.