Improve step sizes for x-axis labels in bins graphs
The labels for the bitloading graph stay relatively unchanged. They are
now dynamic, and always always use power of two steps.

For the other bins graphs, switch to dynamic decimal-based labels, which
should be easier to read. This also fixes an issue where multiple labels
had the same text on ADSL lines due to rounding.