Fix handling of zero/invalid values for showbins data in DrayTek client
Filter out the SNR value -32, as it seems to be sometimes reported in
the output of the showbins command in the place of invalid values
(presumably because the actual value -32.5 is truncated to an integer).

Also make sure to not return uninitialized zero values in the SNR data.
There are two sources for this currently: First, the output from the
device may contain less values than the number of carriers (which the
size of the showbins data slice is based on). Second, the code currently
stops copying at the last valid value when the group size is guessed to
be different from 1 (but the uninitalized target slice may be larger).