Add showtime state to errors history graphs
Periods where the line was not in showtime state are marked with a red
background.

Invalid periods are still determined based on the actual data instead of
using the Valid flag of the showtime state. This way the graph is still
marked gray if no data is available despite showtime state (for example
for retransmission counters when G.INP is not in use).