Fix upstream Hlog scaling for Fritzbox devices on VDSL2 Profile 35b
Normally, the QLN and Hlog data in the support data seem to be reported
in groups as defined by the VDSL2 specification. The group size is not
reported, so it is guessed from the bands data.

However, it looks like AVM applies some scaling to the upstream data
when VDSL2 Profile 35b is in use, so that the guessed group size is
incorrect. As an attempt to fix this, adjust the group size by a factor
of 2 in that case.