Add DrayTek web client
This client uses the web console instead of Telnet for running commands.

It works well in principle, but unfortunately the showbins output is
truncated to 240 entries, and the snr/qln/hlog commands don't return any
output at all.

In addition to that, the device only allows a single active session on
the web interface. Any existing session is automatically terminated.