Improve compatibility of macOS binaries from GitHub workflow
Currently, an older SDK is selected explicitly to achieve compatibility
with older macOS versions. However, it is also possible to achieve this
using the environment variable MACOSX_DEPLOYMENT_TARGET along with a
current SDK.

With this change, Mac OS X 10.15 is supported again, just like it was
before the recent SDK update.