Improve Linux build process in GitHub workflow
The Linux build process is now split into two steps: First, a container
image with all dependencies is built. Then the application is built
using the container.

Some hacks are used in the container image to allow dependencies for
multiple architectures to coexist in a single image.

However, in the end it is worth it, as the build times are reduced by
caching of the container image.