Make line break / carriage return handling in Telnet client more robust
According to RFC854, a carriage return on the same line should be
encoded as CR-NUL. So this needs to be handled when trying to detect
repeated prompts.

Also make the removal of the echoed command more robust by recognizing
any combination of CR/LF/NUL as line break. This is necessary for some
DrayTek devices (e.g. Vigor 167), which use CR-CR-LF as line break.

This also applies for removal of the prompt (last received line).